[add] #171 ヘルプダイアログのdartファイル化
